requirements.txt
asgiref==3.7.2
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
environs==9.5.0
marshmallow==3.19.0
packaging==23.1
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3

pytest~=7.4.0